server/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.7.0
django-heroku==0.3.1
django-ratelimit==3.0.1
djangorestframework==3.12.4
dnspython==1.16.0
gunicorn==20.1.0
idna==2.10
psycopg2==2.8.6
pycodestyle==2.7.0
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
whitenoise==5.2.0
